Start using size_t instead of int in ShapeList

+++ src/ShapeFiles/ShapeList.cpp
@@ -6,7 +6,7 @@ bool ShapeList::hasConflictsWith( shapePointer_t insertShape )
 
   shapePointer_t object1;
 
-  for (int i = 0; i < shapes.size() && conflict == false; i++) 
+  for (size_t i = 0; i < shapes.size() && conflict == false; i++)
   {
     object1 = shapes(i);
     conflict = object1->isTouching( insertShape  );
@@ -14,19 +14,19 @@ bool ShapeList::hasConflictsWith( shapePointer_t insertShape )
   return conflict;
 }
 
-int ShapeList::addShapeToList( shapePointer_t insertShape )
+size_t ShapeList::addShapeToList(shapePointer_t insertShape)
 {
-  int curSize = shapes.size();
+  size_t curSize = shapes.size();
   shapes.resize(curSize + 1);
   shapes(curSize) = insertShape;
   return curSize;
 }
 
-int ShapeList::addList( ShapeList addList )
+size_t ShapeList::addList(ShapeList addList)
 {
   compressed_vector<shapePointer_t> addShapes = addList.getShapes();
-  int additionSize = addShapes.size();
-  int curSize = shapes.size();
+  size_t additionSize = addShapes.size();
+  size_t curSize = shapes.size();
   shapes.resize( curSize + additionSize );
   foreach_( shapePointer_t curShape, addShapes )
   {
@@ -41,12 +41,12 @@ int ShapeList::addList( ShapeList addList )
 int ShapeList::removeShapeFromList( shapePointer_t shapeToRemove )
 {
   compressed_vector<shapePointer_t> newShapeVector;
-  int newSize =  shapes.size();
+  size_t newSize =  shapes.size();
   newShapeVector.resize(newSize);
   bool removedShape = false;
 
 
-  int curIndex = 0;
+  size_t curIndex = 0;
   int removedIndex = -1;
   foreach_( shapePointer_t curShape, shapes)
   {
@@ -62,29 +62,25 @@ int ShapeList::removeShapeFromList( shapePointer_t shapeToRemove )
       removedIndex = curIndex;
     }
   }
-  if ( removedShape )
-  {
+  if ( removedShape ) {
     shapes = compressed_vector<shapePointer_t>( newShapeVector );
   }
 
   return removedIndex;
 }
 
-int ShapeList::clearShapes() {
+size_t ShapeList::clearShapes() {
   shapes.clear();
   shapes.resize(0);
   return shapes.size();
 }
 
-compressed_vector<shapePointer_t> ShapeList::getShapes()
-{
+compressed_vector<shapePointer_t> ShapeList::getShapes() {
   return shapes;
 }
 
-void ShapeList::update(const float dt)
-{
-  foreach_ ( shapePointer_t curShape, shapes )
-  {
+void ShapeList::update(const float dt) {
+  foreach_ ( shapePointer_t curShape, shapes ) {
     curShape->update( dt );
   }
 }
